bgp: Reorganise bgp directory structure
In preparation for introduction of FRR, reorganise the bgp directory
structure.

The "bgp" package will now consist of a sub-package containing
the "native" BGP implementation. Code that is common to all
implementations is pulled out into the "bgp" package.
